Fix compilation error on gcc

+++ tools/remap/source/light_ydnar.cpp
@@ -1169,7 +1169,7 @@ static float DirtForSample( trace_t *trace ){
 	}
 
 	/* apply gain (does this even do much? heh) */
-	outDirt = std::min( 1.0f, pow( gatherDirt / ( numDirtVectors + 1 ), dirtGain ) );
+	outDirt = std::min( 1.0f, powf( gatherDirt / ( numDirtVectors + 1 ), dirtGain ) );
 
 	/* apply scale */
 	outDirt *= dirtScale;
